Implement delete middle and delete duplicate
Function q_delete_mid is used to delete the middle element in
entire queue, use two pointers visit the queue, when fast pointer
point to head or last element, delete the element which slow pointer
point to.
Function q_delete_dup is used to delete duplicate element in
entire queue, use one string to record the element value which
is be visit, and use strcmp to compare if value is same with
value of pre-element or not.